Allow construction using pre-parsed array
Hi,

Perhaps it would be useful to be able to instantiate the class using a pre-parsed array?

I have a use case where I use svg-path-commander to parse a bunch of paths beforehand, which I then later use with this library. (this implementation of getPointAtLength is _so_ much faster than path-commanders :)

Thanks for the good work!